[cm] Allow the built in ELF runner to be routed and used.
Currently, the built-in ELF runner is implicitly used when no other
runner is specified, and there is no way to explicitly use it.

This change offers the ELF runner as a capability to the root component,
allowing it to be routed like any other runner capability.

The longer term goal is to avoid the ELF runner being implicitly made
available to all components in the system, but instead only being
available where it is explicitly offered.

Bug: 4761
Change-Id: I594e298ba8883402ef2b2daa68cf8782688b6d71