[appmgr] Add an allowlist for ambient replace-as-executable
This one is a little different from the other allowlists: we only wish
to enforce it on _user builds, so we add a feature to allowlists to
detect if we actually read a list, or if we failed to find anything.

Then we can include the allowlist in `base_package_labels` in the product
configurations we want to restrict.

To support testing this change, I needed to inject a config directory
handle into appmgr, because it's not straightforward to test different
config-data package contents, since that's a global product
configuration.  So that's an FD now.  And also json_parser needed to
grow the ability to ReadDirectoryAt, so that's in here (with test!) too.

Bug: 37924
Change-Id: Ida4749c0f58adc8ffdddee93adc6101e70840255