[kernel][vm] Remove TA_NO_THREAD_SAFETY_ANALYSIS
This removes most of the usages of TA_NO_THREAD_SAFETY_ANALYSIS in the
VmObject hierarchy and introduces usages of AssertHeld and TA_ASSERT
instead.

The logic of CreateClone is rearranged slightly as, once analysis was
enabled, turned out to be performing actions without appropriate locks.

Test: No behavioral changes. Clang build gave no analysis errors and
running tests in debug build did not result in AssertHeld failing.

Change-Id: I49522cbe9f26eeb98cc8678f45af86b698a12ae4