[graphics][compute] Fix flakiness in vk_app_state_t initialization.
Calling vk_app_state_init() would sometimes fail when running unit
tests on the CQ (see attached bug for details). The reason was
trace_provider_create_with_fdio() failing to register to the trace
registry sometimes, for some unknown reason (!?).

To work around this, only call the function when tracing is required
by the caller.

Another issue fixed by the CL is that the ASSERT_MSG() call after
the trace_provider_create_with_fdio() didn't check for the right
condition. As such, failure was undetected, leading to crashes
in fuchsia_state_destroy() when trying to pass the NULL pointer
to trace_provider_destroy().

Bug: 41918
Test: fx run-test graphics_compute_vk_unittests

Change-Id: Ifcfe0de7228aae399be7bb7e72d6fda462521348