[graphics][compute] Add spinel/spinel_opcodes.h temporary header.
The Spinel styling opcodes definitions are supposed to be an
implementation detail of the Spinel library, but for now their
definitions are essentially required by any test program, which
led to their duplication into several source files.

This CL removes all the duplicates and puts a single definition
in a new header that is included both by the implementation and
the tests.

A warning comment is provided explaining that this will be removed
in the future.

Bug: None
Test: fx run-test graphics_compute_unittests
Change-Id: Ibb648772d76221c3cb29454a9ab5f901ea17fcbd