[kazoo][go] Remap `func` and `g` when used as identifiers
In trying to actually roll the new generation into third_party/go, it
fails because |type|, |func|, |g| are used as argument names in the fidl
syscalls. Go doesn't like this, so rename them to something similar.

Additionally, map some PCI types to uintptr for types that the Go
bindings currently don't define (and aren't likely to be used from Go
anyway).

Test: Updated existing tests, and added new test case specifically for reserved words
Bug: 39918
Change-Id: Ia6dd3280e3e51883b30f23785c617f2ae0a56dd2