[debugger] Parse string literals.
Adds a parser for C and Rust string literals. This should handle regular
and raw strings and all escape sequences except unicode ones.

It does not handle Rust "byte" sequences or C unicode/wide strings.

This is currently not hooked up. In a future pass it will be hooked up
to the expression tokenizer for literal strings and the command line
input.

Change-Id: Icaf26c72deec51bdfec382b66f00edc09407d7b4