[fvm][utest] Make first mount call in CorruptMountHelper synchronous
There has been flakes in TestCorruptMount in which the second mount
call when creating a filesystem leads to an already bound error when
trying to acquire the block device fifo. This change makes the
first mount call synchronous to wait until the mount process that is
supposed to error finishes and releases its connections before continuing.

Bug: 39510
Change-Id: I1a5b233a4c75fe10b2eccfa1306c5442711df8b2