[media] overhaul media bundles/packages files
This is https://fuchsia-review.googlesource.com/c/fuchsia/+/343895,
which was reverted due to a missed reference in boards/vim2.gni.
That reference has been added to this CL.

This CL is a general overhaul of the bundle/package files used in
media. It accomplishes the following:
1) References to //src/media/... are removed from garnet except
   as noted for soft transition
2) A consistent system of bundles/BUILD.gn files is established
   for src/media
3) Exact behavior with respect to generated package lists is
   maintained.
A lot of structure that should be happening in //bundles is instead
happening in //garnet/packages. As a result, some media-related
bundles targets may appear closer to the root of the bundle tree
that one would consider appropriate. Hopefully, this will get fixed
in the future when //bundles matures. I didn't want to take on
reorganizing //bundles itself.

I've verified 3) by running a set of scripts that build all products
(fx list-products) for two boards, one arm64 and one x86.

Change-Id: I914a7ad9d28b2cf937c80552af86f59752b60d74