[kernel] remove generic user callback
Continues the program of fxb/40105 to split the usermode
thread with the core thread. In this change we use the
fact we have a real C++ typed pointer  to the usermode
thread to simplify the callback code.

But more importantly to remove one more unmanaged pointer
that point up the stack.

BUG = 40105

TEST = this is a refactor, no functionality changed.
Some tests touched but those are manual / interactive tests
that still pass via 'k thread_tests'

Change-Id: Ib164beab71858de232be7d427c3673ddf1fa417a