[tools][net][mdns] Updates connection error handling.
This is to handle errors similarly to the netboot library, as can be
seen here:
https://fuchsia.googlesource.com/fuchsia/+/05e0e510b37fd755ed7f9f421ade93a1c5fe612b/tools/net/netboot/netboot.go#170

This will make testing in CI/CQ environments less error prone, as it is
possible that the mDNS library can connect to some interfaces but not
others. In cases where the client fails on encountering the first error,
it may be possible that an error is encountered without attempting to
connect to an interface that would have joined the mcast group or
connected to the IP in question without error.

Change-Id: Ib99dc37e20074797747930918f33f55dc6f82b26