[kernel][ddk] Eagerly Unpin PMTs
The zx_pmt_unpin descriptions claims that it unpins pages that were
pinned. Currently there is a race condition that can cause pages to
remain pinned after the zx_pmt_unpin syscall has returned to user.

The race condition is possible through the ZX_INFO_PROCESS_VMOS of
zx_object_get_info as this syscall will end up in
ProcessDispatcher::ForEachHandleBatched. ForEachHandleBatched will
acquire RefPtr's to the dispatchers for a selection of handles, then
drop the handle table lock and call the provided callback on the
dispatchers.

If the callback happens in parallel with zx_pmt_unpin then the
RemoveHandle call in unpin can succeed (since ForEachHandleBatched
drops the handle table lock), but as there is still a reference to
the dispatcher the PinnedMemoryTokenDispatcher destructor will not
get called at the end of sys_pmt_unpin, and it is through the
destruction of the dispatcher and the associated PinnedVmo, that
causes unpin to actually happen.

This changes MarkUnpinned to Unpin that eagerly causes the undeerlying
PinnedVmObject to be released, instead of delaying it to PMT dtor, that
may happen arbitrarily far in the future.

Test: Added core test that would reliably fail without these changes.

Bug: 36257

Change-Id: Ie4356e53e7dd4298aeccbfdef1fc57b0c8b65b43