[FDF] Call suspend hook with a suspend reason.
Platform devices like acpi implement suspend hook which will be used to
shut down the platform. For system power states like reboot, ACPI writes
a reboot reason to the firmware, before halting the
system. This change adds a way to convey the suspend reason to the
suspend hook along with the device power state the device has to go
into for a particular system power state. The suspend reason is like a
hint for the driver, to understand, why the suspend hook is called. It
is not system power state.

Test: runtests -t ddk-power-test.

Change-Id: I8ae21769c935efe7760c3651ed8d60de8147b174