[camera] Switch to BufferCollectionInfo_2 (step 1)
This CL is one of several that transitions the ISP interface from
using BufferCollectionInfo to BufferCollectionInfo_2 structures.
It adds an implementation of CreateOutputStream2() in preparation for
its addition to isp.banjo.

Testing: Ran camera_tests.

Note that this CL only adds a function that has yet to be called.
Its implementation simply returns an error code in this CL.
It will be fleshed out and invoked in a subsequent CL,fxr/344690.

Change-Id: If8cb0b87cc3de35e01dd0ee2b011717573ccdbdf