[escher/flatland] Flat Rectangle Mesh Generator
Add new functionality to tessellation.h/cc to create
an IndexedTriangleMesh2d of a flat rectangle on the XY
plane given an origin point, a size, and UV coordinates.

This is part of the effort to implement a FlatlandRenderer.

Testing: New unit test in indexed_triangle_mesh_clip_unittest.cc.

Change-Id: Iffe8cd5eea6a1873bdf6b43c4b70760194790481