[fidlgen][llcpp] Fully qualify zx::unowned_channel, zx::channel, and zx::time.
Some locations in generated llcpp bindings refer to types in zx without
a leading ::.

As part of fxb/39732, which is attempting to make the
//zircon/syscalls/zx.fidl a normal standalone fidl file rather than a
fidlc-builtin, there will be a <zx/llcpp/fidl.h> generated.

However, once that exists, there's a conflict between namespace
llcpp::zx and namespace zx (where the definition of of unowned_channel,
etc. can be found). To fix this, fully qualify some references to types
that were being used without the leading ::.

No intended change in functionality.

Bug: 39732
Change-Id: I2345532543b427cbbc6e517b4eaf2792ffd87fbf