[zircon][x64] Fix wrong bits being read for power unit calculations.
Power unit calculations are documented (both in Intel MSR documetation
and in-code comments) to read bits 0 to 3 of the MSR RAPL_POWER_UNIT.
However, the code incorrectly reads bits 0 to 15. Update the code to
match the documentation.

Discovered during testing of ubsan in the kernel, due to the large shift
causing undefined behaviour.

Tests: The system fails quickly in unit tests to a ubsan violation
before this CL, but is able to pass after this CL.

Change-Id: Idd299aa61f3c04d3db55639f3512c0400f29f42d