[fidl] Reland "[fidl] Removes fuchsia.timezone"
Uses that remain need to move first to
fuchsia.deprecatedtimezone.

From there, we'll be deprecating the service by replacing
the current uses by equivalent functionality.  As this
will last a while, let's do the easy thing first.

This change will be submitted *last*, after all dependent uses
are migrated or removed.  It depends on the Dart SDK having been
migrated off of timezone into deprecatedtimezone.

Change-Id: I44a3067aa999acf620b6168e4c24055e363a46cb

---

This reverts commit cbf0756c1d8d1e59ba5376624cedd3bdada38277.

Reason to revert: the changes that needed to precede this one have now
been stitched back up; and using fx make-integration-patch to confirm
that it does not break.

Tested:
  fx make-integration-patch

Bug: 8895
Bug: 46002
Change-Id: Iec12535d2db6fc410b8c1dc8b99c423744407d3e