[escher/flatland] Refactor Renderer Inheritance
Refactor some of the methods in PaperRenderer to
the base Renderer class as they will be useful for
other renderer subtypes, such as the FlatlandRenderer.
Also refactor the "FrameData" struct so that there is
a base FrameData struct that each renderer subtype
inherits from and extends so that common data can
be shared.

Testing: This is a pure refactor with no new functionality
added, so there are no new tests added. All existing tests
continue to pass.

Change-Id: Iaca59b978983ec2c2882f1f6a8334bdf4f57db45