[camera][calibrate] Factory Protocol Impl
This CL adds the scaffolding to the factory calibration API method calls. It allows for a stream from the IspTester to be opened and for several frames to be written to disk. Future CLs will allow us to specify how many frames are to be written and will add on to the rest of the methods specified by the API.

This CL also moves all BLC logic to an "algorithms" directory,

Tested: `factory_protocol_test.cc` added. Several frames able to be
written to disk (scp'd to host and examined manually).

Change-Id: Ib8b7b2f1245239e93491a1a872128beb8c74ec20