[camera] Switch to BufferCollectionInfo_2 (step 2)
This CL is one of several that transitions the ISP interface from
using BufferCollectionInfo to BufferCollectionInfo_2 structures.

This CL adds CreateOutputStream2() to isp.banjo.
CreateOutputStream2() differs from CreateOutputStream() in that it
BufferCollectionInfo_2 and ImageFormat_2 instead of
BufferCollectionInfo.

Note that while all relevant code contains implementations of
CreateOutputStream2(), none of them use it yet.

Testing: Ran the camera test suite.

Note that this CL only adds a call to the isp.banjo interface.
The actual use of this call, along with tests that use it, are
in a subsequent CL,fxr/344690.

Change-Id: Iaacb8a1ee116a0960a01a66a75177239205e26c7