[fidl][media] Disable MediaPlayerTests temporarily
Due to temporary code needed for the FIDL v1 transition, MediaPlayer
tests are taking a significant amount of time to run on CQ/CI and are
causing the arm64-asan bots to time out. This CL disables the tests
temporarily, and will be re-enabled once the performance impact caused
by the FIDL v1 transition can be mitigated.

Test: (this CL disables several tests)
Change-Id: Id2c1b26602f6ca3a66fa1b31714c1f4a4010e9ee