[ledger] PageUpload: remember new commits received in backoff
When in a backoff, PageUpload was ignoring the OnNewCommits signal. This
used to be OK: after backoff, we created a new BatchUpload object with
all available commits. We are now retrying the existing upload: we
should remember that some commits are pending upload but not in the
batch so we can start a new batch afterwards.
This is done by changing the meaning of the PageUploadState: it is now
indicating when a batch is uploading or in backoff. Instead of calling
PreviousState/NextState to retry, we call the "Resume" method. During a
backoff, OnNewCommits calls NextState normally.

Test: PageUploadTest.NewCommitsDuringBackoff fails before

Change-Id: Ib6d26dd76244eca66e02b03772abff9981726bd7