[debugger] Add support for ELF symbols.
Adds the ELF symbols to the list of addresses and symbols returned by
ModuleSymbols. Although most are duplicates of DWARF, this adds some
important things such as names for vtables and RTTI information which
will be used in the future.

Currently these are not used for much but will be needed more in the
future for printing virtual classes.

Change-Id: I1d3d6a6eb90f291eb27d9789c3d0de0738f06668