[fs] The big migration to LLCPP
================ Notes For Reviewers ================

Despite parts being mechanical, this ended up being quite a large CL.
I'll break down the changes here, happy to carve them out into
individual CLs if it helps the review:

- The old <fs/handler.h> FIDL utilities were replaced by an
implementation that works with LLCPP transactions. However, it turns out
that src/devices/coordinator and src/bringup/virtcon are using it. So
the old handler implementation was moved to
`src/storage/deprecated-fs-fidl-handler` instead.

- HandleFsSpecificMessage is still load-bearing, and takes a C binding
`fidl_txn_t`. A `fs::internal::CTransactionShim` class was created to
expose a C binding transaction interface in front of LLCPP. This does
lead to the unfortunate situation of an LLCPP transaction wrapped in a C
shim wrapped in an LLCPP shim in certain cases. A follow up change will
move HandleFsSpecificMessage and all its clients to LLCPP, getting rid
of the shim.

- `VnodeRepresentation::ConvertToIoV1NodeInfo` was introduced to extract
the innards of a `VnodeRepresentation` and invoke a callback with the
pointer to a stack-allocated `NodeInfo`. The use of a callback ensures
that the payloads of `NodeInfo` are guaranteed to live longer than the
accessing code. This prepares it for the upcoming FIDL migration from
unions to xunions.

- Parts of the `Connection` class relevant to FIDL dispatching are
carved out into a `std::shared_ptr<Binding>`. This allows the binding to
be borrowed via `weak_ptr`s by in-flight transactions. In particular, in
the `SyncTeardown` case, the binding could be destroyed independent of
in-flight transactions referencing it. I'm not strongly committed to
this approach. Happy to discuss better designs.

- In order to support vfs_unmount_handle with a deadline, the LLCPP
`fidl::Call` method was extended to plumb a deadline to `zx_channel_call`.

Thanks.

======================== End ========================

Thanks to many preparatory CLs, this is mostly a 1-to-1 mechanical
replacement of C binding types/functions to their LLCPP counterpart.

An important change is a custom LLCPP transaction to support
coorperative connection teardown during vfs shutdown.
This is located at ulib/fs/fidl_transaction.cc

Bug: 34613
Test (fs side): no behavior change (but add a few more test in
teardown_tests)
Test (llcpp timeout): runtests -t fidl-llcpp-interop-test

Change-Id: Ia1774e28e6b84deb0992447e4d69d9c45091a173