[line input] Prompt for options.
Adds a helper to manage asking for one of a set of valid inputs. This
would be used for most common questions like yes/no ones.

In order to support optionally accepting the input without requiring the
user type "enter", an additional callback on the line input is added
when the input has changed.

Hooks up the new call to the zxdb console but does not yet use it for
anything.

Change-Id: I8c9a04ccfa0c7e0cd9b7b7730066cc4a2596bdb3