[media.sessions] WatcherSink generic over Sinks
This CL refactors the WatcherSink type into two parts. Previously,
WatcherSink implemented both back pressure and filtering as a sink along
the way to the FIDL client.

Now, the backpressure is implemented in FlowControlledProxySink, and the
filter is just a function that can be applied to any stream by

  .filter_map(watcher_filter(filter))

This enables a following CL, which creates a channel to receive watcher
events internally so they may be routed to clients of SessionControl
on request.

MULTIPLY: `[{
    "target": "//src/media/sessions:mediasession_tests",
    "total_runs": 100
}]`

Bug: 42166
Test: fx run-test mediasession_tests
Change-Id: Idf7213347ab6cc804eb3577f1835755bf9d430c9