[wlan][mlme] Add logger shim
Currently the Rust logger in the MLME isn't hooked up to anything, as
devmgr doesn't hook its logger up to anything. We try to hook up the
logger ourselves here, but some other Rust library might try to do the
same thing: that's okay, as long as we get some kind of logger in the
end.

Test: Tested manually
Bug: 41417
Change-Id: I3ceab8cc1e0c1e3d35a12b2281e1be8de5e6220c