[kernel][events] Propagate event signal result through early wakeups
The kernel contains an event type which allows a thread (or threads) to
wait for an event to be signaled. When an event is signaled, the
signaler may optionally provide a signal result that can be observed
by the waiter. This mechanism had a race condition. If a thread manages
to start waiting on an already signaled event, it will not observe the
result used by the waiter. This fixes the race condition by storing
the result in the event structure so it can be propagated to waiters
that encounter the early return.

Tests:

This adds tests for the propagation of the result value through kernel
event signaling calls. There are two cases tested here:

*) event_signal_etc() called before any thread blocks on event_wait()
*) event_signal_etc() called after a thread blocks on event_wait()

The first case is tested on a single thread where ordering is easy to
arrange.  The second case is harder to arrange deterministically so the
test makes a best effort to sequence things and gives up if it can't
get the sequence it wants in a reasonable period of time.

BUG: 34521
Change-Id: Ie230df71e18b03a480597d69231626a417b9c92c