[fzl][flake] Avoid flake in VMAR move tests
So, during extended validation of the fix for bug 41331, the vmo_move test
showed failures on 4/24 test machines after 45 hours of focused testing.  The
cause was functionally identical to the flake in the core tests, but _much_ less
likely to happen because of the lack of address space pressure in the move test
when compared to the core tests.

Avoiding this issue is easier.  Simply introducing a sub-vmar to run the test in
will avoid the possibility that a death-test thread's stack mapping ends up
conflicting with the test VMO mapping.

Bug: 41331
Change-Id: I4f02967d9dd6d1698b8bd0ac934e6af25af4f779