[system-ota-test] Fix hostname construction with IPv4 addrs
If the hostname we're constructing is not an IPv4 address, we
shouldn't enclose it in brackets, as that's not a valid URI. The check
for this looks for a colon in the hostname, because hostnames with
colons are invalid per
https://tools.ietf.org/html/rfc3986#section-3.2.2 UNLESS they are IPv6
literals.

Test: Ran the system OTA test in GI using the DOWNLOAD_BLOB code,
which exercises fuchsia's hyper adapter, which previously failed on
the system OTA test.

Bug: 42342
Change-Id: If2d9a0bcd15bb068b288c08f58eea962745bc84d