[identity] Use fuchsia.identity.external.Error in google auth provider
This migrates all the errors in google auth provider
to use the errortype defined in the new set of auth
provider APIs. The auth_provider module still needs
to serve AuthProviderStatus and is left as the only
portion of the binary that handles the old
AuthProviderStatus. A thin conversion is added so
that the auth_provider module can continue to use
the rest of the crate. Once the migration to
fuchsia.identity.external is complete,
AuthProviderStatus and the auth_provider module
will be removed entirely.

Testing: unit tests, google_signin_e2e

Bug: 565
Change-Id: I8e15d9e5e1ad62e437d23b1538820e437364bb60