[debugger] Separate breakpoint execution scope.
BreakpointSettings contained a "scope" for where it applies (target,
thread, or global). This patch separates out that state into its own
class.

This separate class cleans up some of the syntax. In a followup this
will be used more in a cleanup of the "edit" command for breakpoint
settings.

Converts the pointers to WeakPtr which is theoretically unnecessary but
makes the lifetimes of these much harder to mess up.

There should be no behavior change.

Change-Id: I8e58a07d521b3f31cd57543cfe4aae24e9754512