[fidl][rust] Refactor encoding/decoding of nullable types
This CL refactors the way we encode and decode nullable types in Rust.
It replaces the traits Autonull and AutonullContainer with a single,
more flexible Autonull trait. Implementing this trait for T results in
automatic implementations of Encodable/Decodable for Option<Box<T>>,
Option<&mut T>, and Option<OutOfLine<T>>. The latter is temporary and
will soon be removed along with the OutOfLine type itself.

In addition, this CL:

- Fixes fxb/42150, allowing nullable xunions to be used as method
  parameters in protocols.
- Removes code for encoding/decoding nullable tables, since tables are
  not allowed to be nullable in FIDL.
- Changes OutOfLineUnion to be an alias for OutOfLine. This type was a
  workaround introduced for the union-to-xunion migration, and will be
  removed along with OutOfLine in a future CL.

Test: fx run-test rust_fidl_tests
Test: fx build garnet/public/lib/fidl/rust/fidl:fidl.rust.test.compilation
Bug: 42304
Fixed: 42150
Change-Id: Ic5568d5628d53e3655603758bb4d223f6bbfa61c