[build] fix zircon fidl hack
This change modifies the zircon fidl compat hack to use get_path_info
instead of get_label_info when checking if a fidl library is in zircon.
The latter would fail with paths such as //test/foo.fidl as '//' is not
a valid label, but it is a valid path.

Test: verified locally that //experiment/experiment.fidl builds
      CQ
Change-Id: Ie38bd7d36462ff51125d0e08dd7288ec6bac3922