[component-manager] Add component_manager to the testing build dependencies.
In certain builds the component-manager is available as a package, in
others it is in the ZBI. The test cases assume the fuchsia-pkg will be
available when in fact fuchsia-boot may be the only available source for
the package given the build type.

To address this since it is  only an issue in test it makes the most
sense to just always add the component_manager package to the test
build.

FIXES: 42465
Change-Id: Ibb277616a6241403e0caac40a11b5e506a8a6ebb