[testsharder] Check for multipliers based on Name instead of Target.
As a soft transition to using `name` instead of `target`, we kept both
but populated TestModifier.Name with the Target value and set the Target
to "". However, when creating the multiplier shards, we were still
checking the TestModifier.Target instead of Name so it wasn't creating
any multiplier shards since Target was empty.

Change-Id: I02bd36dfad8a9fca3034018587e1f939ee35d0e7