[audio_core] Further decouple Stream from Packet
The |Packet| class is essentially a buffer descriptor that also holds
the FIDL callback to be invoked when the packet is released. These
details are not particularly relevant to the stream interface, which
really just wants to read the next buffer in the stream iteratively.

To decouple these, we define a new 'Buffer' class inside of stream that
holds the general buffer descriptor, but we keep the
fbl::RefPtr<Packet> as an implementation detail of the PacketQueue.

Bug: 36279
Test: fx run-test audio_core_unittests.
Change-Id: I393cfd73a3b01d0503474632eefaa3f66ff6e6f9