[banjo][rust] Update logic to derive Debug and PartialEq
In the case where a struct field is a type identifier, look up
whether that type implements Debug and PartialEq to determine whether
the struct itself should derive Debug and PartialEq. This change is
needed to generate ddk.protocol.wlan.mac.

In the case where struct field is an array and the size type is a
constant identifier, look up the value it holds to determine whether to
derive.

Test:
1. Verified that `fx build ddk.protocol.wlan.mac-rustc` works
2. Added rust-derive.test.banjo
3. `fx exec src/devices/tools/banjo/regen_banjo_tests.sh` to regenerate
   files.
4. fx run-host-tests banjo_unittests

Change-Id: Idec345593f14662e0690c70eeb9e05c51978b62e