[fidl][llcpp] Support for multi-threaded dispatch.
Adds support to the LLCPP server bindings for multi-threaded dispatch,
asynchronous Unbind(), and other related changes.

Doc: go/fidl-mt-server-cpp

The BindingRef API has been modified as follows:

1. OnChannelErrorFn<> and OnChannelCloseFn<> have been replaced with a
general purpose OnUnboundFn<> which is invoked when a channel is unbound
from the dispatcher.

2. Unbind() now initiates an asynchronous unbind operation. On
completion, the user is given ownership of the channel endpoint via the
OnUnboundFn<> if specified. Unbind() may be invoked from any thread.

* NOTE: For a single-threaded dispatcher, in most instances the behavior
of Unbind() is unchanged in that on return, no other entities will access
the channel. The primary exception is in the case that Unbind() is invoked
within the scope of the message handler for the same binding. In that
case, deletion will necessarily be postponed to after the message handler
returns. In all cases, if provided, the OnUnboundFn<> will be executed.

3. The user may now invoke Close() from a BindingRef. It is identical to
Unbind() except that it triggers the sending of an epitaph with the
given status and the subsequent closing of the channel endpoint. The
user cannot recover the channel endpoint.

The CompleterBase API has been modified as follows:
1. The user may now enable concurrency in message handling via the new
EnableNextDispatch() API. It's effect is to add back the async wait for
the binding to the dispatcher before the message handler returns.
Subsequent calls are idempotent, and the wait is not added a second time
after the message handler returns. As this API only has value within the
message handler, it is restricted to the Completer<>::Sync variant.

Bug: 40787,38456,35107

Test: runtests -t fidl-async-test-test.
Change-Id: I7d078a0a2c73f0c568f348a07968f10895fa462e