[identity] Remove GetFirebaseToken API.
This method should not have been included in this API since firebase
tokens may be issued by a different organization than the refresh
token this API requires. The removal here simplifies the ongoing
migration to fuchsia.identity.external.

The only client of the method was Ledger's FirestoreCloudProvider which
was recently removed in fxr/340931. This API is included in the SDK for
legacy reasons but there are no clients outside stem.

Change-Id: I1f29997f565acbd925ab2e1c0cb85bade93d8669