[escher] Use hardcoded pipeline stage and access masks in BatchGpuUploader/Downloader
Since we always have separated command buffers for BatchGpuUploader/Downloader, and
we always use VkSemaphores to synchronize them, we can ensure that, there are
only TransferWrite operations in BatchGpuUploader and there are only TransferRead
operations in BatchGpuDownloader.

Therefore, we can hardcode these pipeline stage and access masks so that we only
make sure Write (or Read) operations happens in correct order when we do image
layout transitions.

TEST=escher_tests, scenic_tests

     Add a test case in BatchGpuUploader test -- we can upload to one same image
     and change its layout multiple times without causing any validation
     error.

Change-Id: I3a693465e7b89096061a774a7f710580608d79cc