[identity] Move mocks to nested mod
This moves the mocks defined within the auth_provider
test mod to a separate mod. This enables usage of the
mocks in multiple test contexts, which will be
necessary to test the multiple protocols in
`fuchsia.identity.external`.

This also adds async-trait to simplify definition of
the traits somewhat.

Testing: unit tests

Bug: 565

Change-Id: I901fbeedfc636a4d3d92477c9d1a42c8fc5e0e0b