[camera] Switch to BufferCollectionInfo_2 (step 4)
This CL is one of several that transitions the ISP interface from
using BufferCollectionInfo to BufferCollectionInfo_2 structures.

This CL removes the CreateOutputStream() from the ISP Banjo
protocol and cleans up code no longer in use.
It is the last step in the sequence of CL's for this change.

Testing: Ran the camera test suite and camera demo.
Change-Id: Id7c268e29479687c098c0cdfba562d3c575f3e60