[zxcrypt] Prepare for build unification
zxcrypt-dev is intended to be used by device drivers, which means it
requires static linkage, since otherwise we can trigger deadlocks in
dynamic linking if the library loader is requesting a library off disk.

zxcrypt-fdio is intended to be used by non-driver programs, and can be
dynamically linked.

Lastly, both of these are going to need to be exposed at zircon/public/lib
to complete build unification, so go ahead and expose them in the SDK to
make that easier later.

Change-Id: I597008da4effa5e7cc9318c78f876fb38c0e4e94