[scenic] Use fit::optional for debug_name of View and ViewHolder
Fixed: 42475

We fixed the bug in command applier where we may dereference a
nullopt when creating View and ViewHolder.

Also we change the debug_name to fit::optional in sdk commands.
Since we use "string?" as types of debug_name in ViewArgs and
ViewHolderArgs in FIDL definition, it would be more appropriate
to use fit::optional in SDK as well.

TEST=gfx_unittests

     added a unit test in view_unittest.cc for creating nullable
     View and ViewHolder.

Change-Id: I7bef160eceace775328294cfd5cdef622d9c8924