[fidlc] Fix unresolved type alias ICE
When compiling a usage of a type alias, it's possible the alias
declaration has not been resolved. Previously, this would result in an
ICE due to attempting to access unresolved size bounds.

This change checks for resolution and resolves the declaration if
necessary before accessing the type's values. The logic is very similar
to that of the type declaration code path.

Fixed: 41932
Test: Added new tests to prevent similar class of bugs.
Test: Run the fidl-compiler-test host test.

Change-Id: I9ea01ed2f10ca3b8dfa21d0b8601a9b37b7cdf5b