[feedback] removed production config
* The production and default config files have not yet diverged and
don't look to be any time soon so there is no need for both to exist.
This should reduce the overall complexity of feedback agent since we
no longer need to maintain fake packages for the sake of testing
config overrides and there any new annotations that are added only need
to be represented in a single file.

TESTED=`fx run-test feedback_agent_tests`
TESTED=`crasher` dedad2a2f2bfd753

Change-Id: I88321162351a4e21a39c52a4d5c22ecae0208310