[pm] Metadata should be in UTC with no nanoseconds
TUF-1.0 section 4.4.2 states that the metadata expiration must be in
ISO-8601 format in the UTC timezone with no nanoseconds. This patch
fixes pm to generate the proper format. At the moment go-tuf does not
enforce the timezone, but rust-tuf does, so this is a blocker for the
rust-tuf migration.

Fixed: 42490

Change-Id: I1005b1a28fc0c55395daf225ef1bc82bd693db76