[FDF][1/4] Move system suspend/resume interface out of device manger
This change adds a new interface for suspending and resuming devices.
It also modifies the mulitple-device-test to use the new interface. As a soft
transition, after all the clients are moved to the new one, we can
remove the old interface. The new interface takes a system power state
as input as compared to the old one which takes in flags. Eventually we
will not need the flags anymore, once the coordinator's suspend routine
takes in SystemPowerState.

Bug:41071

Test: fx run-test coordinator_tests

Change-Id: Ie49cf2cbb36b396c9dad249da020c00f804de121