[zircon] Require AllocChecker be checked if ZX_DEBUG_ASSERT_IMPLEMENTED
(relanding commit f8800ef9db709428077e15e4071a4a8f77476420)

This is part of a series of changes to clean up use of LK_DEBUGLEVEL.

Prior to this change, when AllocChecker was used in non-kernel code,
its panic_if_armed method was a no-op because it was gated on
LK_DEBUGLEVEL, which was never defined outside of kernel code.  This
meant that AllocChecker did not enforce its rule that check() must be
called after each allocation.

1. Change AllocChecker's panic_if_armed feature to be conditional on
ZX_DEBUG_ASSERT_IMPLEMENTED.

2. Fix a few instances where AllocChecker was not being checked.

Bug: 41790
Change-Id: I2f9739c773d89962dd67dddc46fac7a43cb42caa