[session] Add keyboard input
Renames touchable_session to input_session and adds keyboard and mouse
support.

This change does not add any special handling for multiple devices of
the same type. The current design does not propose how this will be
handled, yet, and it is not a priority. However, some development
environments have multiple devices of the same type. For example,
running Fuchsia via AEMU from Chrome Remote Desktop results in multiple
keyboard devices in Fuchsia (the physical keyboard and the virtual,
remote keyboard). In most cases, only one keyboard is active at a time,
but it is unnecessarily complicated to determine which of the keyboard
devices is the active keyboard. Rather than adding special, temporary
code to explicitly determine and connect to only the active keyboard,
this change just listens for events from all connected keyboards,
without any other special handling.

Change-Id: I5a5742a50b3c71ffa60683b28caadd0686806b21