[debugger] Settings can store ExecutionScopes.
This is a step in making breakpoints use a SettingStore for the
breakpoint settings. This will allow us to use "set" for breakpoint
settings so we can delete the very confusing "edit" command.

Converts SettingValue to a class and make the members protected.

This is not currently hooked up to any actual settings so there should
be no visible behavior change.

Change-Id: Idc5b47161b7fd21740f32f97481319dc0d688e34