[chromium][cast] Move system configuration files to src/{chromium,cast}
This moves the system configuration data files for the Chromium and
Cast components from the central {appmgr,sysmgr} directories to the
src/{chromium,cast}/config directories, close to the rest of the build
and configuration files for these components. This makes it easier to
reconfigure these components as a single unit.

This also adds support for the 'visibility' attribute in the
config_data GN template to make it safer to use.

Bug: crbug.com/1029180, 39800
Change-Id: Ib0c5f0c7af54598c9910d688487b172e95aa88ed