[process_builder] Ignore reservation VMAR drop failure
We're seeing a component_manager panic several times a day in CI/CQ.  We
believe it's attributable to dropping the process handle on some other
failure before dropping the reservation VMAR handle; a theory which is
supported by a core test mimicking that behavior.

This change makes us ignore errors when dropping the reservation VMAR --
we haven't seen any failures here on normal success paths, and these
panics are covering up error propagation for other failures.  This
should help us get CI/CQ unflaked while we try to gather more
information on the actual failures that are triggering this failure, and
to buy time to add additional tests on the process_builder side.

Bug: 42437
Change-Id: Ifb34c19767677604e1c28b7a3529071188150b6f