[wlan][sme] Intersect rates in AP
This seems like as good of a choice as any other to handle conflicting
rates from the client, and should be more resilient to bogus clients.

This also breaks out errors from intersect_rates, and also adds a
separate reason_code field to AssociationError so they don't have to be
implicitly inferred from the result_code.

Test: Unit tests
Bug: 37891
Change-Id: I63361fdddba133ef669bc8bae4160ecfb385fdf1