[benchmarks] Split SL4F perf test wrapper into multiple test executables
Split out from perf_test.dart 3 files for the separate tests.

This is beneficial because the Fuchsia Infra tools for retrying tests
and tracking flakes operate on the granularity of test executables.

For example, after this change, if the Ledger perf tests flake, they
will get retried without retrying the other perf tests, and
flake-fetcher will file a separate bug on "ledger_perf_test".

Unfortunately this change makes it harder to run the full set of
SL4F-based perf tests locally: there is no way to do that without
manually listing the 3 test wrappers.  However, it is now easier to
run each of the test wrappers individually.

Note: I would have preferred to use the name "ledger_perf_tests"
(plural) to reflect that this is a set of tests, but the filename
filter in build/dart/group_tests.py does not allow that.

Bug: 23091
Bug: 39577
Bug: 42576
Test: on CQ: terminal-x64-release trybot;
  local testing: "fx set terminal.x64" + e.g. "fx run-e2e-tests microbenchmarks_test"
Change-Id: Ie942e693ce083b93eb7f4fc8a6664606d686ce76