[fidl] fidl_transform manually sent FIDL messages
This CL adds a WriteTransform method to fidl::Message, which
is used to transform manually sent messages to v1. This ensures
that all unions are xunions on the wire in preparation for
flipping unions to xunions in fidlgen.

Only messages that contain unions are updated to minimize
the amount of work needed (and since the risk that these
messages are updated to contain unions between this CL
and the union/xunion flip is small).

Test: update all bindings to only read v1, and run that
in GI

Change-Id: Ia265be9a1d98e57f894365b4f1c6a00eabf45cea