[magma] Tear down MagmaSystemDevice in unbind
The unbind hook is intended to set the device state so that clients
close their connections and stop issuing queries. For magma that
involves tearing down the MagmaSystemDevice, which will close all the
channels to clients.

Without this, the shutdown test would often (especially in debug builds)
keep sending GPU commands for over 1 second after unbind without
receiving an error and accidentally keep the old GPU driver open too
long. When the new GPU was bound on the same piece of hardware that
would cause a number of errors and make the test fail.

Fixed: 42286

Change-Id: I4092a4476fb31f0e50584f1897b704a9a7ed609b