[fs] Assert that Vfs::SetDispatcher can only be called once
It seems a bit dangerous that the dispatcher will be swapped out in the
presence of in-flight transactions. An alternative is to always require
the dispatcher when creating a Vfs, but it turned out to be quite
involved as the memfs API creates a Vfs but does not take in a
dispatcher.

Change-Id: I14e2bb656b46f1b015098abeff6adf9215e49b21