[pkgfs] Add /pkgfs wrapper library
pkgfs is used by both the package cache and the package resolver, and
the bindings to interact with it using fuchsia.io are a bit verbose.
This change introduces a new rust crate to expose the /pkgfs filesystem
via typesafe wrapper types and refactors known instances of interacting
directly with pkgfs to use these wrapper types.

With the following exception, no functional changes are expected:
* The FileProxy drop close guard synchronously sent the file close
request, but with the recently added blob fetch retries, it became
more likely for the retry to attempt to re-open the blob before pkgfs
processed the close request. These new wrappers require clients to wait
for the close operation to complete before dropping the file proxy.

Test: some basic new test cases for the pkgfs library, as well as
regression coverage from //garnet/tests/pkg-resolver-tests

Change-Id: Ied69cf87172edb197b715a484cc3cc1a9e46dc2d