[camera] Switch to hw_accel_frame_callback (step 1)
This CL is one of several that transitions the ISP interface from
using OutputStreamCallback to using the hw_accel_frame_callback
that is common to other camera-related hardware accelerators.
The hw_accel_frame_callback provides additional information about
a frame, including its metadata and status. The changes also move the
name of CreateOutputStream2 back to its original CreateOutputStream.

This CL adds an implementation of CreateOutputStream() in preparation for
its addition to isp.banjo.

Testing: Re-ran all the camera tests, as this CL only adds a function
that has yet to be called.

Change-Id: Ibaa655c76899625d77b49dfa8dc8f6736228a1a6