[audio_core] Simplify Usage handling
  * For Renderers and Capturers, remove the 'allowed_usages' list. This
    is currently always the full set of usages and the FIDL bindings
    will validate the incoming value is one audio_core was built
    against.
  * For Renderers only, we forbid usage changes once the stream type
    has been configured. Capturers continue to allow runtime usage
    changes since this is a load bearing feature used to implement
    system agent audio ducking.

The reason for disallowing mutating Renderer usage at runtime is that
as part of enabling a more complex mix pipeline (involving multiple
mixer passes), we will start to route these renderers to different
mixer instances based on the assigned usage. While we could surely take
on the complexity of re-routing these streams in response to a Renderer
usage change, we would much rather clients simply create multiple
AudioRenderers to output audio with different usages instead of
changing the assigned usage of a single Renderer.

Bug: 36279
Test: GI dry-run
Change-Id: I1f7fdfb98f4aab4a3eb3b9bbbfa179f93a11df1d