[benchmarks][channel] Microbenchmark for zx_channel_write_etc and zx_channel_read_etc
The existing benchmark for zx_channel_write and zx_channel_read was
updated to write and read varying numbers of handles. The reason is
that the *_etc and non-etc syscalls should have mostly identical
behavior when handles are not written.

Note: the names of the 0-handle test cases get renamed by this change.

Bug: 41920

Change-Id: I3c0ebe48cb6eb797c88062a0235e9aad74a8f5be