[escher] Make Waterfall demo testable.
A major part of the refactoring is to break the bidirectional
dependency between Demo <-> DemoHarness.  Now, DemoHarness refers
to Demo, but not vice versa.  This makes it easier to embed a Demo
in a unit test, without complications due to the third-party GLFW
library.

Some feature changes were made, motivated by design decisions that
made along the way.  For example, WaterfallDemo::stopwatch_ was used
for both animating lighting as well as FPS computations.  When moving
this into DemoHarness, the decision was made to leave a stopwatch in
WaterfallDemo for lighting animation.  Now that it's not the same one
as for FPS computation, an existing keyboard control was modified to
allow lighting animation to be paused.

Bug: 42436

Change-Id: I8984df0082ab07dca4426c31dda3fa6725638c4f